wip: add preparing state for team

+++ crates/web/src/router/api/game/game_id/team/team_id/mod.rs
@@ -1,14 +1,14 @@
+mod avatar;
+mod token;
+mod user;
+
 use axum::{Router, http::StatusCode};
 use cds_db::{
     entity::{team::State, user::Group},
     get_db,
     transfer::Team,
 };
-use sea_orm::{
-    ActiveModelTrait,
-    ActiveValue::{Set, Unchanged},
-    ColumnTrait, Condition, EntityTrait, NotSet, QueryFilter,
-};
+use sea_orm::{ActiveModelTrait, ActiveValue::{Set, Unchanged}, ColumnTrait, Condition, EntityTrait, NotSet, PaginatorTrait, QueryFilter};
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 
@@ -17,14 +17,11 @@ use crate::{
     traits::{Ext, WebError, WebResponse},
 };
 
-mod avatar;
-mod token;
-mod user;
-
 pub fn router() -> Router {
     Router::new()
         .route("/", axum::routing::put(update_team))
         .route("/", axum::routing::delete(delete_team))
+        .route("/ready", axum::routing::post(set_team_ready))
         .route("/state", axum::routing::put(update_team_state))
         .nest("/avatar", avatar::router())
         .nest("/users", user::router())
@@ -44,7 +41,7 @@ pub struct UpdateTeamRequest {
 /// # Prerequisite
 /// - Operator is admin or one of current team.
 pub async fn update_team(
-    Extension(ext): Extension<Ext>, Path((game_id, team_id)): Path<(i64, i64)>,
+    Extension(ext): Extension<Ext>, Path((_game_id, team_id)): Path<(i64, i64)>,
     Json(mut body): Json<UpdateTeamRequest>,
 ) -> Result<WebResponse<Team>, WebError> {
     let operator = ext.operator.ok_or(WebError::Unauthorized(json!("")))?;
@@ -52,12 +49,7 @@ pub async fn update_team(
         return Err(WebError::Forbidden(json!("")));
     }
 
-    let team = cds_db::entity::team::Entity::find()
-        .filter(
-            Condition::all()
-                .add(cds_db::entity::team::Column::GameId.eq(game_id))
-                .add(cds_db::entity::team::Column::Id.eq(team_id)),
-        )
+    let team = cds_db::entity::team::Entity::find_by_id(team_id)
         .one(get_db())
         .await?
         .ok_or(WebError::BadRequest(json!("team_not_found")))?;
@@ -83,22 +75,26 @@ pub async fn update_team(
 }
 
 pub async fn delete_team(
-    Extension(ext): Extension<Ext>, Path((game_id, team_id)): Path<(i64, i64)>,
+    Extension(ext): Extension<Ext>, Path((_game_id, team_id)): Path<(i64, i64)>,
 ) -> Result<WebResponse<()>, WebError> {
     let operator = ext.operator.ok_or(WebError::Unauthorized(json!("")))?;
     if operator.group != Group::Admin {
         return Err(WebError::Forbidden(json!("")));
     }
 
-    let team = cds_db::entity::team::Entity::find()
-        .filter(
-            Condition::all()
-                .add(cds_db::entity::team::Column::GameId.eq(game_id))
-                .add(cds_db::entity::team::Column::Id.eq(team_id)),
-        )
+    let team = cds_db::entity::team::Entity::find_by_id(team_id)
         .one(get_db())
         .await?
         .ok_or(WebError::BadRequest(json!("team_not_found")))?;
+
+    if team.state != State::Preparing {
+        return Err(WebError::BadRequest(json!("team_not_preparing")));
+    }
+
+    let _ = cds_db::entity::team_user::Entity::delete_many()
+        .filter(cds_db::entity::team_user::Column::TeamId.eq(team.id))
+        .exec(get_db())
+        .await?;
 
     let _ = cds_db::entity::team::Entity::delete_many()
         .filter(cds_db::entity::team::Column::GameId.eq(team.game_id))
@@ -108,6 +104,65 @@ pub async fn delete_team(
 
     Ok(WebResponse {
         code: StatusCode::OK,
+        ..Default::default()
+    })
+}
+
+/// Set a team's state to Pending.
+///
+/// # Prerequisite
+/// - Operator is admin or one of the members of current team.
+pub async fn set_team_ready(
+    Extension(ext): Extension<Ext>, Path((game_id, team_id)): Path<(i64, i64)>,
+) -> Result<WebResponse<Team>, WebError> {
+    let operator = ext.operator.ok_or(WebError::Unauthorized(json!("")))?;
+
+    if operator.group != Group::Admin
+        && !cds_db::util::is_user_in_team(operator.id, team_id).await?
+    {
+        return Err(WebError::Forbidden(json!("")));
+    }
+
+    let game = cds_db::entity::game::Entity::find_by_id(game_id).one(get_db()).await?
+        .map(|game| cds_db::transfer::Game::from(game))
+        .ok_or(WebError::BadRequest(json!("game_not_found")))?;
+
+    let team = cds_db::entity::team::Entity::find_by_id(team_id)
+        .one(get_db())
+        .await?
+        .map(|team| cds_db::transfer::Team::from(team))
+        .ok_or(WebError::BadRequest(json!("team_not_found")))?;
+
+    if team.state != State::Preparing {
+        return Err(WebError::BadRequest(json!("team_not_preparing")));
+    }
+
+    // Review the number of members
+    let team_users = cds_db::entity::team_user::Entity::find()
+        .filter(cds_db::entity::team_user::Column::TeamId.eq(team.id))
+        .count(get_db()).await?;
+
+    if team_users < game.member_limit_min as u64 || team_users > game.member_limit_max as u64 {
+        return Err(WebError::BadRequest(json!("member_limit_not_satisfied")));
+    }
+
+    let team = cds_db::entity::team::ActiveModel {
+        id: Unchanged(team.id),
+        game_id: Unchanged(team.game_id),
+        state: Set(if game.is_public {
+            State::Passed
+        } else {
+            State::Pending
+        }),
+        ..Default::default()
+    }
+    .update(get_db())
+    .await?;
+    let team = cds_db::transfer::Team::from(team);
+
+    Ok(WebResponse {
+        code: StatusCode::OK,
+        data: Some(team),
         ..Default::default()
     })
 }
